+++ src/main/java/hashing/NativeHashCodeComputer.java
@@ -10,7 +10,7 @@ public class NativeHashCodeComputer extends  HashCodeComputer {
     }
 
     @Override
-    protected int hashCode(byte[] input, Unsafe unsafeAccess, long address, long off, long length) {
+    protected int hashCode(byte[] input, Unsafe unsafeAccess, long address, int off, int length) {
         long hash = 0;
         int i = 0;
         for (i = 0; i < length; i += 4) {
